添加钱包后，新的钱包数据集会同步显示到主窗口

+++ Utils/IconManager.cs
@@ -12,7 +12,8 @@ namespace AccountBook.Utils {
             return newIcon;
         }
 
-        public static void DeleteIcon(string icon) {
+        public static void DeleteIcon(string? icon) {
+            if (icon == null) return;
             var fullPath = FullPath(icon);
             File.Delete(fullPath);
         }
